requirements.txt
Django==2.2.14
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
